Renamed BaseVariableWithEvent<T, TEvent> to BaseVariable<T, TEvent>
