--- Assets/SO Architecture/Editor/Inspectors/BaseVariableWithEventEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 99ba96ff2f6b66146ba89808c09398a1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
