--- Assets/SO Architecture/Variables/BaseVariableWithEvent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fa6d6ee583d04dd4b87f62a3251f2f84
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
